Make duplicate tcplink and ptlink objects
This commit exists to make reviewing this PR easier.
This commit may be squashed with several of the following
commits before merging,

The PluggableTransports client and link are based on the TCP client and
link. I am copying the files because they are largely the same code, and
the next commit will contain the changes.

I am not subclassing because The common TCP link object knows too much
about asio sockets, and the PluggableTransports link does not rely on a
system socket being available. This can possibly be solved by a
refactor, but I think it's out of scope for this Pull Request.

Signed-off-by: Nick Beirne <nick@tunnelbear.com>