Move link and cli to the PluggableTransports namespace
This is mostly renaming functions, changing the namespace, and renaming
log functions. I am also editing logs to reflect PT and not TCP.

During this rename I also needed to make ptlink stand alone, and not
reliant on TCP link base. This are only two virtual functions which need
to be implemented.

I also needed to add pluggable transports errors to help accomidate the
new transport.

Signed-off-by: Nick Beirne <nick@tunnelbear.com>