Switch to PluggableTransports::Transport in ptcli
Since we can't guarantee async operations exists for a given pluggable
transport we now rely on asio's post operation. As a side-effect, errors
are also no longer from asio, and instead are expected to be an
ExceptionCode (but it falls back if an exception is thrown and its not
that type).

In async_send and async_recv we explicitly don't capture self. This is
because the post happens in a different thread, and there would be a
race condition between reference counted pointers to self. Instead we
rely on the caller to maintain a strong reference to self for the
duration of the operation.

Signed-off-by: Nick Beirne <nick@tunnelbear.com>