ProtoContext: added keepalive_timeout_early parameter
keepalive_timeout_early defines the keepalive_timeout
parameter early in the connection before the KeyContext
reaches ACTIVE.

It is set via the optional third parameter to the
"keepalive" directive, for example:

  keepalive 1 8 4

sets keepalive_timeout_early to 4 seconds.  If unspecified,
keepalive_timeout_early defaults to keepalive_timeout.

keepalive_timeout_early is useful on the server side to
reduce the resource footprint of abandoned connections,
and can be tuned to mitigate DDoS and UDP amplification
attacks.

Signed-off-by: James Yonan <james@openvpn.net>