fix reset of auth-token on AUTH_FAILED,SESSION
This is e.g. sent during renegotiation when the lifetime of an auth-token
on the server has expired.

Due to shortcut evaluation of the logical OR expression the call to
can_retry_auth_with_cached_password() never happened with connections
that use autologin(_session). Since the called function has side effects
on the ClientCreds object, i.e. resetting the session_id to the saved
password, an endless loop of retries with the expired auth-token
followed.

Signed-off-by: Heiko Hund <heiko@openvpn.net>