vcpkg-ports: remove openssl port and use the latest version from repo
We want to upgrade to OpenSSL 1.1.1l, but instead of upgrading
our local port we can now rely on the OpenSSL version shipped
with the vcpkg's upstream repo.

We can now do so because starting from vcpkg's commit
06f8fd63 ("[OpenSSL] support for "no-autoload-config" config option
(#18389)") support for the OpenSSL no-autoload-config build option
has been added upstream. This option is required by our builds.

Until now this option was hardcoded in our custom port, but now we
can turn it on by defining our custom triplets "<arch>-windows-ovpn3".

Given the above, we can now drop the custom port and directly rely
on the upstream repo shipping the most recent OpenSSL version.

Signed-off-by: Lev Stipakov <lev@openvpn.net>