ClientConnect: try to re-cache stale RemoteList items
Try to resolve and cache addresses again each time the tunnel is
re-negotiated. Only items which are beyond their lifetime are considered.

This is done hoping that the VPN DNS is able to resolve server addresses
correctly, so that those --remotes which have changed addresses can be
used again in the future.

The motivation for this change were slowly but steadly outdated pre-cached
remotes when --persist-tun is used. Eventually all cached addresses are
outdated, resulting in loss of connection.

While this approach is not guaranteed to work it's probably better than
doing nothing.

Signed-off-by: Heiko Hund <heiko@openvpn.net>