ClientProto: add client_proto_renegotiated callback
Add a NotifyCallback for tunnel renegotiation. It is called whenever
a secondary KeyContext reaches the active state, i.e. is fully negotiated.

The already existing virtual active() method is extended to have a bool
parameter to signal if it was the primary (i.e. first) or a secondary
KeyContext which was negotiated, so that ClientProto can call the
callback then.

Signed-off-by: Heiko Hund <heiko@openvpn.net>