remove RemoteList::first_item() method
first_item() returns a nullptr in case there's no first item in the
RemoteList. This behavior can be achieved by calling defined() before
get_item(0). In addition to that existing code failed to check for
nullptr and dereferenced whatever was returned. Thus the method was
considered harmful and is removed.

Signed-off-by: Heiko Hund <heiko@openvpn.net>