RemoteList: don't ignore item in use in BulkResolve
Remove the exception for the item in use when re-resolving a RemoteList.
This can be done now, that Item are referred to with a smart pointer
instead of a direct reference from extern.

Instead reset the address index for that item, so that it doesn't point
to a potentially non-existent address.

Signed-off-by: Heiko Hund <heiko@openvpn.net>