[OSSL1.1/FIPS] Ignore unsusable cipehrs when running with FIPS mode
When running with the old FIPS mode in OpenSSL 1.0.2 and the Red Hat
patched OpenSSL 1.1.1 RHEL8 we need to check if the ciphers are actually
usable. Otherwise we get nasty errors when trying to use them later.

Note that OPenSSL 1.0.2 will still fail when the server picks
the OpenVPN PRF as we do not have a FIPS compatible PRF function for
OpenSSL 1.0.2.

Signed-off-by: Arne Schwabe <arne@openvpn.net>