OVPN3-800 build: re-enable -flto for linux deps
lz4 and mbedtls are currently the only deps
that are built for linux (asio and xxHash
are only copied, not built).

If LTO flag is specified (which is a variable
already supported by scrips/build) and target
is linux then enable -flto but allow to
speficy a random-seed to allow reproducibility.