Fix ASAN detected new/delete mismatches
The unique_ptr_slab variation of the std::unique_ptr<T> addresses the
issue of new/delete mismatches in code that allocates a _memory slab_
with the global _operator_ new but de-allocates an _object_ with a
delete _expression_.  The use case that manifests the mismatch is as
follows: Allocate a slab of memory that has a C struct at the head of
the slab, with a "my_type mt[0];" as the head's last member.  The slab
is cast to the type of the C struct, but sized to contain N my_type
items.

The object based de-allocation is the behavior of the
std::default_delete<T> template; it is used by the std::unique_ptr<T>
if the user does not specify an alternative deleter.  The
unique_ptr_slab resolves the mismatch with an alternative deleter that
de-allocates the _memory slab_ with the global _operator_ delete.

Signed-off-by: Mark Deric <jmark@openvpn.net>