endian64: Check of __GLIBC__ availability
When compiling this code using the musl libc instead of glibc, it fails
with this error:

./openvpn3-core/openvpn/common/endian64.hpp: In function 'uint64_t openvpn::Endian::rev64(uint64_t)':
./openvpn3-core/openvpn/common/endian64.hpp:53:14: error: '__bswap_constant_64' was not declared in this scope
   53 |       return __bswap_constant_64(value);
      |              ^~~~~~~~~~~~~~~~~~~

The __bswap_constant_64() is a function provided by the glibc library
and is not available in all other libc implementations.  To avoid this,
we fallback to the same solution used for Clang, which builds fine.  But
to avoid missing a match on the MINGW32 or MSC environments, the #if
condition checks are slightly reordered.

Signed-off-by: David Sommerseth <davids@openvpn.net>