openvpn/http: document source and add license text
While the link is probably fine for most people
it is not strictly enough to fulfill the license
terms. Since this code is very limited in scope
I decided to add the txt file not in top directory
but rather near the files where it matters.

Add PURL to document where the source came from.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>