omi: add missing line ending
As a response for correct password, omiclient
must send SUCCESS: with \r\n line ending.

This bug went unnoticed for a while because
the only user of this code as for now is openvpn-gui
with openvpn3 support, which hasn't got much
traction until recently.

Missing line ending didn't trigger certain callback
in -gui code which enables logging, and logging
was enabled in a different way (one can call it
openvpn3-specific hack). With that change a -gui callback,
which enables logging, is called and hack could be removed.

Signed-off-by: Lev Stipakov <lev@openvpn.net>