Add code style configurations
This adds a clang-format with the new code style for OpenVPN 3.

A very light Emacs dir-locals configuration has also been added, but
that will most likely not cover all the formatting aspects which
clang-format does.  Running clang-format before committing is still
highly recommended to ensure the coding style is consistent.

Signed-off-by: David Sommerseth <davids@openvpn.net>