Ensure sending UV_ only with push-peer-info and prioritise client.peerInfo
This commit changes the way the core deals with UV_ variables. They
now also respect push-peer-info (like in OpenVPN 2.x) and if variables
are present in both client.peerInfo and as setenv in the configuration
content, only the ones from peerInfo are send to the server.

The new behaviour can be tested with

	ovpncli -I UV_TEST=cmdline conf.ovpn

and conf.ovpn having a setenv UV_TEST foobar in it as well

Signed-off-by: Arne Schwabe <arne@openvpn.net>