Do not allow SHA1 cipher suites when using preferred tls-cert-profile
We do not allow SHA1 in other instances using this profile and while
SHA1 is still fine as HMAC in these situation, people freak out when
seeing SHA1 and also the description and documentation will state
that SHA1 is not allowed in other context (certificate signature),
causing confusion. So better not allow it in this context as well.

Signed-off-by: Arne Schwabe <arne@openvpn.net>