Simplify client options classes
The current passing of client options is very confusing since it is
three classes that largely have the same attributes and the code is
copying them around between the classes.

Instead create a new base class that hold the settings that can be shared
between the classes and only parse/copy the settings that need special
handling. This might keep an extra copy of some settings around but in
constract greatly reduces the code complexity of the options handling.

Also move the check_dco_compatibility function to client options to
be able to avoid carrying around the dco_compatiblity flag.

Signed-off-by: Arne Schwabe <arne@openvpn.net>