Probably reference count the EPKI object and ensure right order
In order to ensure that the xkey provider certificate objects do not
use a dangling pointer, use a shared_ptr to manage the lifetime of
ExternalPKIImpl. Since we need to pass it into a raw void* pointer
we need this raw pointer to a shared_ptr magic here.

Also ensure the right order of epki and config on destruction.

Signed-off-by: Arne Schwabe <arne@openvpn.net>