unittest: Add proto test cmake parameters w/README
When the test/ssl/proto was migrated to a unit test, all the
build time tuneables for changing the test scope got lost.

This change re-introduces these parameters in the appropriate
CMakeLists.txt.  Further, it adds an improved README in .rst format
that describes the parameters; this is based on the original README
file in test/ssl.

In addition, it adds improved handling of the directory where the test
certificates and key files are located; now they can be adjusted more
easily.

Signed-off-by: Mark Deric <jmark@openvpn.net>