Use inline to ensure only one copy of crypto::algs is present
We need to ensure that there is only one copy of these algorithms defined
as we modify them with the method allow_default_dc_algs to set the
F_DC_ALLOW flag on. Having more than one copy means that we have different
copies with different flags which we want to avoid.

The use of inline to a normal variable is a C++17 feature.

From https://en.cppreference.com/w/cpp/language/inline

An inline function or variable (since C++17) with external linkage
(e.g. not declared static) has the following additional properties:

There may be more than one definition of an inline function or variable
(since C++17) in the program as long as each definition appears in a
different translation unit and (for non-static inline functions and
variables (since C++17)) all definitions are identical. For example,
an inline function or an inline variable (since C++17) may be defined
in a header file that is included in multiple source files.

Signed-off-by: Arne Schwabe <arne@openvpn.net>