Fix session_id_time calculation bug
The intent of the offset argument to the calculate_session_id_hmac()
function was to allow the server to accept hmac values calculated for
the previous time interval even as the server's clock moved into a new
time interval.

The bug was that the implementation checked for a match with the
server's current time interval and its _future_ time interval, and not
it's immediate past time interval.

This bug was pointed out during code review by Frank Lichtenheld.  The
code review subject was numeric conversions, in this case signed vs
unsigned. This author changed the offset to unsigned, arguing that the
code would never need to check the client hmac cookie based on both
sides of the server's current time.  The check only needed to be in
one direction, hence unsigned was the solution to the conversion
warning.  That was when Frank pointed out that implementation with the
"+ offset" snippet checked to see if the client provided a psid from
the future time interval and what was needed was to check for the past
interval, i.e., "- offset".

Hence, the change implements the unsigned offset, change add to
subtract from the time range computed for the server's currrent time
window.

Signed-off-by: Mark Deric <jmark@openvpn.net>